refactor: auth module optimizations
