+++ package.json
@@ -75,7 +75,7 @@
       "json",
       "ts"
     ],
-    "rootDir": "src/invoices",
+    "rootDir": "src",
     "testRegex": ".*\\.spec\\.ts$",
     "transform": {
       "^.+\\.(t|j)s$": "@swc/jest"
