+++ src/auth/application/strategies/google.strategy.spec.ts
@@ -1,11 +1,9 @@
 import { Test, TestingModule } from '@nestjs/testing';
-import { PassportStrategy } from '@nestjs/passport';
 import { ConfigService } from '@nestjs/config';
-import { Strategy } from 'passport-google-oauth20';
 
 import { GoogleAuthStrategy } from './google.strategy';
 import { UserService } from '../../../users/application/services/user.service';
-import { mockUserService } from '../../../users/domain/entities/mocks/user-providers.mock';
+import { mockUserService } from '../../../users/domain/mocks/user-providers.mock';
 
 describe('GoogleAuthStrategy', () => {
   let strategy: GoogleAuthStrategy;
